Library Project update5

+++ src/main/java/mft/Controller/BookController.java
@@ -2,7 +2,7 @@ package mft.Controller;
 
 import lombok.extern.log4j.Log4j;
 import mft.model.entity.Book;
-import mft.model.entity.Borrow;
+import mft.model.entity.Management;
 import mft.model.service.BookService;
 
 import java.util.List;
@@ -30,7 +30,7 @@ public class BookController {
                             .build();
             System.out.println(book);
             BookService.getService().save(book);
-            log.info("Save");
+            log.info("save");
             return book;
         } else {
             log.error("Error Save");
@@ -39,7 +39,7 @@ public class BookController {
     }
 
     public List<Book> findAll() throws Exception {
-        log.info("Save");
+        log.info("findAll");
         return BookService.getService().findAll();
     }
 
@@ -62,21 +62,30 @@ public class BookController {
         }
     }
 
-    public Book remove1(String nameBook, String authorBook) throws Exception {
-        Book book = BookService.getService().findByNameBookAndAuthorBook(nameBook,authorBook);
-        BookService.getService().remove1(nameBook,authorBook);
-        log.info("remove");
-        return book;
+    public Book remove1(String username,String nameBook, String authorBook) {
+        try {Management management = ManagementController.getController().findByUsername(
+                    username);
+            if (nameBook != null && authorBook != null && management!=null) {
+                Book book = BookService.getService().findByNameBookAndAuthorBook(nameBook, authorBook);
+                BookService.getService().remove1(nameBook, authorBook);
+                log.info("remove1");
+                return book;
+            }
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+        return null;
+    }
+        public List<Book> searchForNameBook(String nameBook) throws Exception {
+       log.info("searchForNameBook");
+        return BookService.getService().searchForNameBook(nameBook);
     }
 
-    public List<Book> findByNameBook(String nameBook) throws Exception {
-        return BookService.getService().findByNameBook(nameBook);
-    }
 
     public Book findByNameBookAndAuthorBook(String nameBook, String authorBook) throws Exception {
         Book book = BookService.getService().findByNameBookAndAuthorBook(nameBook, authorBook);
         if (book != null) {
-            log.info("save");
+            log.info("findByNameBookAndAuthorBook");
             return book;
         }
         return book;
@@ -84,6 +93,7 @@ public class BookController {
     public Book remove2(Integer id) throws Exception {
         Book book = BookService.getService().findById(id);
         BookService.getService().remove2(id);
+        log.info("remove2");
         return book;
     }
